Dialog de Confirmação

+++ src/app/courses/containers/courses/courses.component.ts
@@ -6,6 +6,7 @@ import { Observable, catchError, of } from 'rxjs';
 import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/error-dialog.component';
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -53,17 +54,25 @@ export class CoursesComponent implements OnInit {
   }
 
   onDelete(course: Course){
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.refresh();
-        this._snackBar.open('Curso removido com sucesso!', '',
-        {
-          duration: 5 * 1000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        });
-      },
-      error => { this.onError('Erro ao tentar remover curso.')}
-    )
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: "Tem certeza que deseja remover esse curso?",
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.refresh();
+            this._snackBar.open('Curso removido com sucesso!', '',
+            {
+              duration: 5 * 1000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          },
+          () => { this.onError('Erro ao tentar remover curso.')}
+        )
+      }
+    });
   }
 }
